Add multi-version support for applications
