+++ Node/README.md
@@ -6,7 +6,7 @@ Ubiq-Genie is a framework that enables you to build server-assisted collaborativ
 
 - **Unity Scenes** serve as the interface for VR users and contain application-specific Unity components that communicate with a server-side `ApplicationController` through a TCP connection, using Ubiq's `Networking` components. These client-side components are written in C# and ensure that outgoing and incoming data are processed and routed correctly. The Unity scenes are distributed as importable samples in the `com.ucl.ubiq-genie` package (see the [package README](../Unity/Assets/Ubiq-Genie/README.md)).
 
-- **Applications** should have an associated Unity scene and `ApplicationController`. The `ApplicationController` is responsible for initializing and managing the services that are required by the application. It also handles the communication between the services and the Unity scene. The `ApplicationController` is written in TypeScript (ESM) and runs on the server. The `ApplicationController` of each of the sample applications can be found in the `app.ts` file in the corresponding folder in the `Node/apps` folder.
+- **Applications** should have an associated Unity scene and `ApplicationController`. The `ApplicationController` is responsible for initializing and managing the services that are required by the application. It also handles the communication between the services and the Unity scene. The `ApplicationController` is written in TypeScript (ESM) and runs on the server. The `ApplicationController` of each sample can be found in `app.ts` inside its app folder (or inside a version subfolder when an app has multiple versions) in `Node/apps`.
 
 - **Services** are modular and can be reused in different applications. Each service is responsible for a specific task and is managed by a `ServiceController`. Services use **providers** — lightweight configuration objects that define what backend to run and how to manage its lifecycle. For instance, the `ImageGenerationService` uses a Stable Diffusion provider that spawns a Python child process to generate images. Providers can be selected via config.json without changing any code — see [Config-Driven Provider Selection](#config-driven-provider-selection). The `ServiceController` is written in TypeScript (ESM) and runs on the server. The `ServiceController` of each of the sample services can be found in the `service.ts` file in the corresponding folder in the `Node/services` folder.
 
@@ -127,6 +127,14 @@ You are now ready to use your new service in an application. For more informatio
 To define a new application, follow these steps:
 
 1. Duplicate the `Node/apps/base` folder and rename it to the name of your application (e.g., `my_application`). Also replace the class name `BaseApplication` in the `app.ts` file with the name of your application (e.g., `MyApplication`).
+
+You can start applications with:
+
+```bash
+npm start <app-name> [version] [configure]
+```
+
+If an app has multiple version subfolders, each version is a direct child folder containing its own `config.json` and `app.ts`. Running `npm start <app-name>` will prompt you to choose a version.
 
 > [!NOTE]
 > The `BaseApplication` application provides a minimal example of creating an application using the `BaseService` service. For more advanced examples, see the existing applications in the `Node/apps` folder. The `registerComponents` method defines the components of the application, which are stored in a dictionary called `components`. The `definePipeline` method defines the pipeline of the application. The `start` method starts the application by registering the components, defining the pipeline, and joining a room on the specified Ubiq server in the configuration file.
