+++ Unity/Packages/com.ucl.ubiq-genie/README.md
@@ -34,6 +34,9 @@ Import samples from **Window → Package Manager → Ubiq-Genie → Samples**:
 
 Each sample has a corresponding Node.js server app (listed in the **Server App** column) located in `Node/apps/` of the main repository.
 
+> [!NOTE]
+> Node apps can define multiple versions as subfolders under `Node/apps/<app-name>/`, where each version folder contains its own `app.ts` and `config.json`. Use `npm start <app-name>` to select a version interactively, or `npm start <app-name> <version>` to start one directly.
+
 ## Package Structure
 
 ```
